decoupling registers
