+++ power/__init__.py
@@ -1,4 +1,4 @@
-from .compute_power import compute_power, power_bounds
+from .compute_power import compute_power
 from .plot import plot_bounds
 
-__all__ = ["compute_power", "power_bounds", "plot_bounds"]
+__all__ = ["compute_power"]
