+++ power/true_effects.py
@@ -1,0 +1,6 @@
+import catalogue
+import numpy as np
+
+true_effects = catalogue.create("power", "true_effects")
+
+@effects.register("effect::cohens_g")
